feat: Duke Level-1
